refactor: ルーティングやミドルウェアの外出し
#19
