+++ web_basic/understand_cors/task_2/pair_quiz/pair_quiz_1/app.js
@@ -1,7 +1,7 @@
 const express = require("express")
 const app = express()
 const apiApp = express()
-const cors = require("cors")
+const router = require("./routes/router")
 const PORT = 8090
 const OTHER_PORT = 8080
 
@@ -16,54 +16,7 @@ app.listen(PORT, () => {
 apiApp.use(express.urlencoded({ extended: true }))
 apiApp.use(express.json())
 
-// CORSの設定
-const corsOptionsDelegate = function (req, callback) {
-  const origin = req.headers.origin
-  const ALLOW_ORIGINS = ["http://localhost:8090", "http://localhost:8091"]
-  const ALLOW_METHODS = ["PUT", "POST"]
-  const ALLOW_HEADERS = ["Content-Type"]
-  const OPTIONS_SUCCESS_STATUS = 204
-
-  // ALLOW_ORIGINSの要素以外のオリジンでは、origin:falseとなるため、CORSエラーが発生する
-  const isAllowOrigin = () => {
-    if (ALLOW_ORIGINS.includes(origin)) {
-      return true
-    }
-    return false
-  }
-
-  const corsOptions = {
-    origin: isAllowOrigin(),
-    methods: ALLOW_METHODS.join(","),
-    allowedHeaders: ALLOW_HEADERS.join(","),
-    optionsSuccessStatus: OPTIONS_SUCCESS_STATUS,
-  }
-
-  console.log(corsOptions)
-  callback(null, corsOptions)
-}
-
-function commonProcess(req, res) {
-  console.log(req.method)
-  console.log(req.params.id)
-
-  const id = req.params.id
-  const resBody = {
-    message: `取得したID：${id}`,
-  }
-  res.json(resBody)
-}
-
-// ルーティングの実装は省略しました...!
-apiApp.options("/users/:id", cors(corsOptionsDelegate))
-
-apiApp.post("/users/:id", cors(corsOptionsDelegate), (req, res) => {
-  commonProcess(req, res)
-})
-
-apiApp.put("/users/:id", cors(corsOptionsDelegate), (req, res) => {
-  commonProcess(req, res)
-})
+apiApp.use("/", router)
 
 apiApp.listen(OTHER_PORT, () => {
   console.log(`listening on port ${OTHER_PORT}`)
